Use common location for common lib and Swagger download.
